Move desktop content tests to features
