+++ test/banners/english/components/BannerCtrl.spec.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it, vi, test } from 'vitest';
+import { describe, test } from 'vitest';
 import { mount, VueWrapper } from '@vue/test-utils';
 import Banner from '../../../../banners/english/components/BannerCtrl.vue';
 import { BannerStates } from '@src/components/BannerConductor/StateMachine/BannerStates';
@@ -8,6 +8,7 @@ import { formItems } from '@test/banners/formItems';
 import { CurrencyEn } from '@src/utils/DynamicContent/formatters/CurrencyEn';
 import { softCloseFeatures } from '@test/features/SoftCloseDesktop';
 import { useOfFundsFeatures } from '@test/features/UseOfFunds';
+import { desktopContentFeatures } from '@test/features/DesktopContent';
 
 const translator = ( key: string ): string => key;
 
@@ -43,38 +44,13 @@ describe( 'BannerCtrl.vue', () => {
 	};
 
 	describe( 'Content', () => {
-		it( 'Plays the slider when the banner state becomes visible', async () => {
-			const wrapper = getWrapper();
-			await wrapper.setProps( { bannerState: BannerStates.Visible } );
-
-			expect( wrapper.find( '.wmde-banner-slider--playing' ).exists() ).toBeTruthy();
-		} );
-
-		it( 'Stops the slider when the form is interacted with', async () => {
-			vi.useFakeTimers();
-
-			const wrapper = getWrapper();
-			await wrapper.setProps( { bannerState: BannerStates.Visible } );
-			await wrapper.find( '.wmde-banner-form' ).trigger( 'click' );
-			await vi.runOnlyPendingTimers();
-
-			expect( wrapper.find( '.wmde-banner-slider--stopped' ).exists() ).toBeTruthy();
-
-			vi.restoreAllMocks();
-		} );
-
-		it( 'Shows the slider on small sizes', async () => {
-			Object.defineProperty( window, 'innerWidth', { writable: true, configurable: true, value: 1300 } );
-			const wrapper = getWrapper();
-
-			expect( wrapper.find( '.wmde-banner-slider' ).exists() ).toBeTruthy();
-		} );
-
-		it( 'Shows the message on large sizes', async () => {
-			Object.defineProperty( window, 'innerWidth', { writable: true, configurable: true, value: 1301 } );
-			const wrapper = getWrapper();
-
-			expect( wrapper.find( '.wmde-banner-message' ).exists() ).toBeTruthy();
+		test.each( [
+			[ 'expectSlideShowPlaysWhenBecomesVisible' ],
+			[ 'expectSlideShowStopsOnFormInteraction' ],
+			[ 'expectShowsSlideShowOnSmallSizes' ],
+			[ 'expectShowsMessageOnSmallSizes' ]
+		] )( '%s', async ( testName: string ) => {
+			await desktopContentFeatures[ testName ]( getWrapper );
 		} );
 	} );
