+++ test/banners/pad/components/BannerCtrl.spec.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it, vi, test } from 'vitest';
+import { describe, expect, it, test } from 'vitest';
 import { mount, VueWrapper } from '@vue/test-utils';
 import Banner from '../../../../banners/pad/components/BannerCtrl.vue';
 import { BannerStates } from '@src/components/BannerConductor/StateMachine/BannerStates';
@@ -9,6 +9,7 @@ import { CloseSources } from '@src/tracking/CloseSources';
 import { CurrencyEn } from '@src/utils/DynamicContent/formatters/CurrencyEn';
 import { softCloseFeatures } from '@test/features/SoftCloseDesktop';
 import { useOfFundsFeatures } from '@test/features/UseOfFunds';
+import { desktopContentFeatures } from '@test/features/DesktopContent';
 
 const translator = ( key: string ): string => key;
 
@@ -44,24 +45,11 @@ describe( 'BannerCtrl.vue', () => {
 	};
 
 	describe( 'Content', () => {
-		it( 'Plays the slider when the banner state becomes visible', async () => {
-			const wrapper = getWrapper();
-			await wrapper.setProps( { bannerState: BannerStates.Visible } );
-
-			expect( wrapper.find( '.wmde-banner-slider--playing' ).exists() ).toBeTruthy();
-		} );
-
-		it( 'Stops the slider when the form is interacted with', async () => {
-			vi.useFakeTimers();
-
-			const wrapper = getWrapper();
-			await wrapper.setProps( { bannerState: BannerStates.Visible } );
-			await wrapper.find( '.wmde-banner-form' ).trigger( 'click' );
-			await vi.runOnlyPendingTimers();
-
-			expect( wrapper.find( '.wmde-banner-slider--stopped' ).exists() ).toBeTruthy();
-
-			vi.restoreAllMocks();
+		test.each( [
+			[ 'expectSlideShowPlaysWhenBecomesVisible' ],
+			[ 'expectSlideShowStopsOnFormInteraction' ]
+		] )( '%s', async ( testName: string ) => {
+			await desktopContentFeatures[ testName ]( getWrapper );
 		} );
 	} );
